@@ -2,14 +2,12 @@ import 'package:mobile/http_wrapper.dart';
 import 'package:mobile/io_wrapper.dart';
 import 'package:mobile/package_info_wrapper.dart';
 import 'package:mobile/preferences_manager.dart';
-import 'package:mobile/database/sqlite_data_manager.dart';
+import 'package:mobile/database/data_manager.dart';
 import 'package:mobile/properties_manager.dart';
 
 import 'device_info_wrapper.dart';
 
 class AppManager {
-  SQLiteDataManager? _dataManager;
-  PreferencesManager? _preferencesManager;
   PropertiesManager? _propertiesManager;
 
   DeviceInfoWrapper? _deviceInfoWrapper;
@@ -17,15 +15,9 @@ class AppManager {
   PackageInfoWrapper? _packageInfoWrapper;
   HttpWrapper? _httpWrapper;
 
-  SQLiteDataManager get dataManager {
-    _dataManager ??= SQLiteDataManager();
-    return _dataManager!;
-  }
+  DataManager get dataManager => DataManager.get;
 
-  PreferencesManager get preferencesManager {
-    _preferencesManager ??= PreferencesManager();
-    return _preferencesManager!;
-  }
+  PreferencesManager get preferencesManager => PreferencesManager.get;
 
   PropertiesManager get propertiesManager {
     _propertiesManager ??= PropertiesManager();
