@@ -6,7 +6,7 @@ import 'package:mobile/res/dimen.dart';
 import 'package:mobile/utils/page_utils.dart';
 import 'package:mobile/widgets/button.dart';
 import 'package:mobile/widgets/list_item.dart';
-import 'package:mobile/widgets/page.dart' as p;
+import 'package:mobile/widgets/my_page.dart' as p;
 import 'package:mobile/widgets/text.dart';
 import 'package:mobile/widgets/widget.dart';
 
@@ -221,8 +221,8 @@ class _ListPickerPageState<T> extends State<_ListPickerPage<T>> {
     List<ListPickerItem<T>> items =
         (widget.allItem == null ? [] : [widget.allItem!])..addAll(widget.items);
 
-    return p.Page(
-      appBarStyle: p.PageAppBarStyle(
+    return p.MyPage(
+      appBarStyle: p.MyPageAppBarStyle(
         title: widget.pageTitle,
         actions: widget.allowsMultiSelect
             ? [
