@@ -1,17 +1,20 @@
 import 'package:device_info_plus/device_info_plus.dart';
 import 'package:mobile/app_manager.dart';
-import 'package:mobile/database/sqlite_data_manager.dart';
+import 'package:mobile/database/data_manager.dart';
 import 'package:mobile/device_info_wrapper.dart';
 import 'package:mobile/http_wrapper.dart';
 import 'package:mobile/io_wrapper.dart';
+import 'package:mobile/model/activity.dart';
+import 'package:mobile/model/session.dart';
 import 'package:mobile/package_info_wrapper.dart';
 import 'package:mobile/preferences_manager.dart';
 import 'package:mobile/properties_manager.dart';
 import 'package:mockito/annotations.dart';
 import 'package:sqflite/sqflite.dart';
 
+@GenerateMocks([Activity])
 @GenerateMocks([AppManager])
-@GenerateMocks([SQLiteDataManager])
+@GenerateMocks([DataManager])
 @GenerateMocks([PreferencesManager])
 @GenerateMocks([PropertiesManager])
 @GenerateMocks([DeviceInfoWrapper])
@@ -23,4 +26,5 @@ import 'package:sqflite/sqflite.dart';
 @GenerateMocks([AndroidBuildVersion])
 @GenerateMocks([AndroidDeviceInfo])
 @GenerateMocks([IosDeviceInfo])
+@GenerateMocks([Session])
 void main() {}
